Deka

+++ app/Http/Controllers/Main/IpbController.php
@@ -83,12 +83,12 @@ class IpbController extends Controller
             })
             ->addColumn('action', function ($data) {
                 $btn = '<div class="d-flex">';
-                // if (auth()->user()->hasPermissionTo('fuel-edit')) {
+                if (auth()->user()->hasPermissionTo('fuel-ipb-edit')) {
                 $btn .= '<a href="javascript:void(0);" class="btn btn-primary btn-sm me-1" title="Edit Data" onclick="editData(\'' . $data->id . '\')"><i class="ti ti-pencil"></i></a>';
-                // }
-                // if (auth()->user()->hasPermissionTo('fuel-delete')) {
+                }
+                if (auth()->user()->hasPermissionTo('fuel-ipb-delete')) {
                 $btn .= '<a href="javascript:void(0);" class="btn btn-danger btn-sm" title="Hapus Data" onclick="deleteData(\'' . $data->id . '\')"><i class="ti ti-trash"></i></a>';
-                // }
+                }
                 $btn .= '</div>';
 
                 return $btn;
